apply Catalog API Blocks and Limit data type from Double to BigDecimal
based on comment on this issue: https://github.com/killbill/killbill/issues/1469#issuecomment-1028580177 .

Related Pull Request in API: https://github.com/killbill/killbill-api/pull/84

+++ catalog/src/main/java/org/killbill/billing/catalog/DefaultTieredBlockPriceOverride.java
@@ -24,14 +24,14 @@ import org.killbill.billing.catalog.api.TieredBlockPriceOverride;
 
 public class DefaultTieredBlockPriceOverride extends DefaultBlockPriceOverride implements TieredBlockPriceOverride {
 
-    private Double max;
+    private BigDecimal max;
 
     @Override
-    public Double getMax() {
+    public BigDecimal getMax() {
         return max;
     }
 
-    public DefaultTieredBlockPriceOverride(String unitName, Double size, BigDecimal price, Currency currency, Double max) {
+    public DefaultTieredBlockPriceOverride(String unitName, BigDecimal size, BigDecimal price, Currency currency, BigDecimal max) {
         super(unitName, size, price, currency);
         this.max = max;
     }
